Rewrite prebuild script into typescript
