+++ package.json
@@ -14,7 +14,7 @@
   },
   "scripts": {
     "dev": "next dev",
-    "prebuild": "node scripts/prebuild.mjs",
+    "prebuild": "jiti scripts/prebuild.ts",
     "build": "next build",
     "start": "next start",
     "lint": "next lint",
@@ -42,12 +42,15 @@
   },
   "devDependencies": {
     "@next/env": "^13.2.4",
+    "@typescript-eslint/parser": "^5.56.0",
     "autoprefixer": "^10.4.14",
     "eslint": "^8.36.0",
     "eslint-config-next": "^13.2.4",
+    "jiti": "^1.18.2",
     "next-sitemap": "^4.0.5",
     "postcss": "^8.4.21",
-    "tailwindcss": "^3.3.1"
+    "tailwindcss": "^3.3.1",
+    "typescript": "5.0.2"
   },
   "bugs": {
     "url": "https://github.com/craigary/nobelium/issues",
